+++ packages/components/src/ellipsis-menu/index.js
@@ -11,6 +11,18 @@ import PropTypes from 'prop-types';
  * This is a dropdown menu hidden behind a vertical ellipsis icon. When clicked, the inner MenuItems are displayed.
  */
 class EllipsisMenu extends Component {
+	constructor( props ) {
+		super( props );
+
+		this.onToggle = this.onToggle.bind( this );
+	}
+	onToggle( isOpen ) {
+		const { onToggle } = this.props;
+
+		if ( onToggle ) {
+			onToggle( isOpen );
+		}
+	}
 	render() {
 		const { children, label } = this.props;
 		if ( ! children ) {
@@ -42,6 +54,7 @@ class EllipsisMenu extends Component {
 				<Dropdown
 					contentClassName="woocommerce-ellipsis-menu__popover"
 					position="bottom left"
+					onToggle={ this.onToggle }
 					renderToggle={ renderToggle }
 					renderContent={ renderContent }
 				/>
@@ -59,6 +72,10 @@ EllipsisMenu.propTypes = {
 	 * A list of `MenuTitle`/`MenuItem` components
 	 */
 	children: PropTypes.node,
+	/**
+	 * Function called when the menu is opened or closed.
+	 */
+	onToggle: PropTypes.func,
 };
 
 export default EllipsisMenu;